Fix the hardcoding of user role using sdk_adapter approach
1. Refer to nova's latest code, use get_sdk to create the client[1].

2. Move placement_client define from default to seperate placement file.

Note:
The new method for get nova and placement client has been tested in latest
cyborg devstack environment. And the test results are good.

Because now cyborg does not have unit tests for the common and conf directory,
unit tests related to this patch will be added in the future, but not in this patch.

Reference:
[1]. https://github.com/openstack/nova/blob/3f019f2db373d82cbc93f0a30de0059d3d07432b/nova/utils.py#L1013

Change-Id: I2a335d47ee3cc24f78b4d49942f1cccd586a0dcc